getting rid of unnecessary loop
